made HighAvailabilityManager abstract
